requirement.txt
fastapi
uvicorn
python-pptx
langchain
langchain-community
langchain-chroma
zhipuai
edge-tts
python-multipart
pydantic
chromadb